+++ backend/src/main/java/com/itsmartsystems/bookyourseat/dto/InvitationRequest.java
@@ -4,25 +4,25 @@ import java.time.LocalDateTime;
 
 public class InvitationRequest {
 
-    private Long senderId;
-    private Long receiverId;
+    private Integer senderId;
+    private Integer receiverId;
     private Long seatId;
     private LocalDateTime startDateTime;
     private LocalDateTime endDateTime;
     
-    public Long getSenderId() {
+    public Integer getSenderId() {
         return senderId;
     }
 
-    public void setSenderId(Long senderId) {
+    public void setSenderId(Integer senderId) {
         this.senderId = senderId;
     }
 
-    public Long getReceiverId() {
+    public Integer getReceiverId() {
         return receiverId;
     }
 
-    public void setReceiverId(Long receiverId) {
+    public void setReceiverId(Integer receiverId) {
         this.receiverId = receiverId;
     }
