+++ backend/src/main/java/com/itsmartsystems/bookyourseat/repository/PostgresUserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface PostgresUserRepository extends JpaRepository<PostgresUser, Long> {
+public interface PostgresUserRepository extends JpaRepository<PostgresUser, Integer> {
 
     Optional<PostgresUser> findByEmail(String email);
 
@@ -14,7 +14,7 @@ public interface PostgresUserRepository extends JpaRepository<PostgresUser, Long
 
     List<PostgresUser> findByRole(String role);
 
-    List<PostgresUser> findByDepartmentId(Long departmentId);
+    List<PostgresUser> findByDepartmentId(Integer departmentId);
 
     Optional<PostgresUser> findByMongoUserId(String id);
 }
